exercise_3
The following exercise assessed my ability to do the following:

Design a logical solution using control structures.
Use relational and logical operators in a programming solution.
Use selection structures in a programming solution.
Use repetition structures in a programming solution.

This program is a "guessing game" program. The program will generate a random integer between 1 and 10000, inclusive. The program should prompt the user to enter a guess. For each guess, the program will output 'HIGHER' if the user's guess is lower than the target, 'LOWER' if the user's guess is higher than the target, or 'WINNER' if the user guesses the target. Each time the program prompts the user for a new guess, it should calculate and display the eligible range of values.