exercise_4
Write a java program for an organization that wishes to accept contributions until a total of $10,000,000 is met. Once this total is hit, no further contributions should be accepted. The organization wants the program to read data from an input file (input.in). The following data should be written to a file called results.out:

The total number of contributions needed to meet the goal of 10 million dollars.
The amount of the largest and smallest contribution accepted.
The average contribution size. Consider both the scenario where the input file does not contain enough contributions to meet the goal and the scenario where the input file contains more data than needed.